fbx fix casts

+++ src/fbx/fbx_geometry.cpp
@@ -17,7 +17,7 @@ namespace FBX {
 		if (data.size() % 2 != 0) throw Exception("data length not a multiple of 2");
 		result.reserve(data.size() / 2);
 		for (size_t i = 0; i < data.size(); i += 2) {
-			result.push_back(OpenGL::UVCoord(data[i], data[i+1]));
+			result.push_back(OpenGL::UVCoord((float) data[i], (float) data[i + 1]));
 		}
 	}
 
@@ -25,7 +25,7 @@ namespace FBX {
 		if (data.size() % 3 != 0) throw Exception("data length not a multiple of 3");
 		result.reserve(data.size() / 3);
 		for (size_t i = 0; i < data.size(); i += 3) {
-			result.push_back(OpenGL::Vector3(data[i], data[i+1], data[i+2]));
+			result.push_back(OpenGL::Vector3((float) data[i], (float) data[i + 1], (float) data[i + 2]));
 		}
 	}
 
@@ -149,7 +149,7 @@ namespace FBX {
 				if (cps.size() % 3 != 0) throw Exception("vertices length not a multiple of 3");
 				controlpoints.reserve(cps.size() / 3);
 				for (size_t i = 0; i < cps.size(); i += 3) {
-					controlpoints.push_back(OpenGL::Vector3(cps[i], cps[i+1], cps[i+2]));
+					controlpoints.push_back(OpenGL::Vector3((float) cps[i], (float) cps[i + 1], (float) cps[i + 2]));
 				}
 				cp_variants_ndx.resize(controlpoints.size());
 			}
